make files independent (#8267)
